python-server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.13.0b2
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
googlesearch-python==1.2.5
huggingface-hub==0.26.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keybert==0.8.5
MarkupSafe==3.0.2
multidict==6.1.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
Pygments==2.18.0
PyYAML==6.0.2
regex==2024.4.16
requests==2.32.3
rich==13.9.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
sentencepiece==0.2.0
soupsieve==2.6
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.5
transformers==4.40.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.16.0